move elastic_tabstops into ext to automatically include in build
