+++ lib/ace/ext/elastic_tabstops_lite.js
@@ -145,14 +145,13 @@ var ElasticTabstopsLite = function(editor) {
 
     this.$setBlockCellWidthsToMax = function(cellWidths) {
         var startingNewBlock = true, blockStartRow, blockEndRow, maxWidth;
-
         var columnInfo = this.$izip_longest(cellWidths);
 
         for (var c = 0, l = columnInfo.length; c < l; c++) {
             var column = columnInfo[c];
             if (!column.push) {
-                console.error(column)
-                continue
+                console.error(column);
+                continue;
             }
             // add an extra None to the end so that the end of the column automatically
             // finishes a block
@@ -253,7 +252,7 @@ var ElasticTabstopsLite = function(editor) {
     // the is a (naive) Python port--but works for these purposes
     this.$izip_longest = function(iterables) {
         if (!iterables[0])
-            return console.error(iterables);
+            return [];
         var longest = iterables[0].length;
         var iterablesLength = iterables.length;
 
@@ -298,8 +297,8 @@ var ElasticTabstopsLite = function(editor) {
 
 exports.ElasticTabstopsLite = ElasticTabstopsLite;
 
-var Editor = require("./editor").Editor;
-require("./config").defineOptions(Editor.prototype, "editor", {
+var Editor = require("../editor").Editor;
+require("../config").defineOptions(Editor.prototype, "editor", {
     useElasticTabstops: {
         set: function(val) {
             if (val) {
