+++ lib/ace/ext/searchbox.js
@@ -43,12 +43,12 @@ dom.importCssString(searchboxCss, "ace_searchbox");
 var html = '<div class="ace_search right">\
     <button action="hide" class="ace_searchbtn_close"></button>\
     <div class="ace_search_form">\
-        <input class="ace_search_field" placeholder="Search for" spellcheck="false">\
+        <input class="ace_search_field" placeholder="Search for" spellcheck="false"></input>\
         <button action="findNext" class="ace_searchbtn next"></button>\
         <button action="findPrev" class="ace_searchbtn prev"></button>\
     </div>\
     <div class="ace_replace_form">\
-        <input class="ace_search_field" placeholder="Replace with" spellcheck="false">\
+        <input class="ace_search_field" placeholder="Replace with" spellcheck="false"></input>\
         <button action="replace" class="ace_replacebtn">Replace</button>\
         <button action="replaceAll" class="ace_replacebtn">All</button>\
     </div>\
